@@ -5,24 +5,36 @@
   "proxy": "http://localhost:4000",
   "homepage": "/react/endless",
   "dependencies": {
-    "axios": "0.18.0",
-    "lodash": "4.17.10",
-    "moment": "2.22.2",
-    "node-sass-chokidar": "1.3.3",
-    "npm-run-all": "4.1.3",
-    "prop-types": "15.6.2",
-    "react": "16.4.1",
-    "react-dom": "16.4.1",
-    "react-router-dom": "4.3.1",
-    "react-scripts": "1.1.4"
+    "axios": "0.19.2",
+    "lodash": "4.17.15",
+    "moment": "2.24.0",
+    "node-sass-chokidar": "1.4.0",
+    "npm-run-all": "4.1.5",
+    "prop-types": "15.7.2",
+    "react": "16.12.0",
+    "react-dom": "16.12.0",
+    "react-router-dom": "5.1.2",
+    "react-scripts": "3.3.0"
   },
   "scripts": {
-    "build-css": "node-sass-chokidar src/ -o src/",
-    "watch-css": "npm run build-css && node-sass-chokidar src/ -o src/ --watch --recursive",
-    "start-js": "react-scripts start",
-    "start": "npm-run-all -p watch-css start-js",
-    "build": "npm run build-css && react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "ncu": "ncu --packageFile ./package.json",
+    "clean-start": "rm -rf ./node_modules/ && ncu -u --packageFile ./package.json && npm i && npm start",
+    "clean-update": "rm -rf ./node_modules/ && ncu -u --packageFile ./package.json && npm i",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
