@@ -1,47 +1,34 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import './Author.css';
+import './Author.scss';
 
 import Profile from './Profile/Profile';
 import StoryGroups from './Stories/StoryGroups';
 
-class AuthorContainer extends Component {
-  constructor(props) {
-    super(props)
+const AuthorContainer = (props) => {
+  const [author, setAuthor] = useState(null);
 
-    this.state = {
-      author: null
-    }
-
+  useEffect(() => {
     axios.get(`api/author/${props.match.params.id}.json`)
       .then(response => {
-        this.setState({
-          author: response.data
-        })
+        setAuthor(response.data);
       })
       .catch(error => {
         console.log(error, 'needs universal error handling');
       });
-  }
-  render() {
-    const {
-      author
-    } = this.state;
+  });
 
-    return (
-      <div>
-        {author && <h3>{author.name}</h3>}
-        {author && <Profile author={author} />}
-        <hr />
-        {author && <StoryGroups stories={author.stories} />}
-      </div>
-    );
-  }
-}
+  return author && (<>
+    <h3>{author.name}</h3>
+    <Profile author={author} />
+    <hr />
+    <StoryGroups stories={author.stories} />
+  </>);
+};
 
 AuthorContainer.propType = {
   match: PropTypes.object.isRequired
-}
+};
 
 export default AuthorContainer;
