@@ -1,9 +1,9 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect }  from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
 import { Route  } from 'react-router-dom';
 import { bcc } from '../Utility/Utilities';
-import './Content.css';
+import './Content.scss';
 import ErrorHandling from '../Utility/Error/Error';
 import Intro from './Intro/Intro';
 import IndexesContainer from './Indexes/IndexesContainer';
@@ -13,74 +13,57 @@ import LibraryCardContianer from './LibraryCard/LibraryCardContainer';
 
 import Story from './Stories/Story/Story';
 
-class ContentContainer extends Component {
-  constructor(props) {
-    super(props);
+const ContentContainer = ({ error }) => {
+  const [allLoaded, setAllLoaded] = useState(false);
+  const [indexes, setIndexes] = useState(null);
+  const [intro, setIntro] = useState(null);
+  const [genres, setGenres] = useState(null);
+  const [story, setStory] = useState(null);
+  const [authors, setAuthors] = useState(null);
+  const [libraryCard, setLibraryCard] = useState(null);
+  const requests = [
+    axios.get('api/intro.json'),
+    axios.get('api/intro.json'),
+    axios.get('api/genres.json'),
+    axios.get('api/stories.json'),
+    axios.get('api/authors.json'),
+    axios.get('api/libraryCard.json')
+  ];
 
-    this.state = {
-      indexesData: null,
-      introData: null,
-      genresData: null,
-      storyData: null,
-      authorsData: null,
-      libraryCardData: null
-    }
-
-    const requests = [
-      axios.get('api/intro.json'),
-      axios.get('api/intro.json'),
-      axios.get('api/genres.json'),
-      axios.get('api/stories.json'),
-      axios.get('api/authors.json'),
-      axios.get('api/libraryCard.json')
-    ]
+  useEffect(() => {
+    if (allLoaded) return;
 
     axios.all(requests)
       .then(axios.spread((
-        intro, 
-        indexes, 
-        genres, 
-        story, 
+        intro,
+        indexes,
+        genres,
+        story,
         authors,
         libraryCard
       ) => {
-        this.setState(prevState => (
-          {
-            ...prevState,
-            introData: intro.data,
-            indexesData: indexes.data,
-            genresData: genres.data,
-            storyData: story.data,
-            authorsData: authors.data,
-            libraryCardData: libraryCard.data
-          }
-        ))
+        setIndexes(indexes.data);
+        setIntro(intro.data);
+        setGenres(genres.data);
+        setStory(story.data);
+        setAuthors(authors.data);
+        setLibraryCard(libraryCard.data);
+        setAllLoaded(true);
       }));
-    
-    this.handleDataReady = this.handleDataReady.bind(this);
-  }
+  }, [allLoaded]);
 
-  handleDataReady(dataId, props = null) {
-    let template = <span>Bad dataId</span>;
-
-    if (this.state[`${dataId}Data`]) {
-      switch (dataId) {
-        case 'intro':
-          return <Intro data={this.state.introData} bbc={this.handleBBC} />;
-        case 'indexes':
-          return <IndexesContainer />;
-        case 'story':
-          return <StoriesContainer genres={this.state.genresData} stories={this.state.storyData} authorsData={this.state.authorsData} />;
-        case 'submit':
-          return <div>SUBMIT</div>;
-        case 'libraryCard':
-          return <LibraryCardContianer />;
-        default:
-          return template;
-      }
-    }
-
+  const handleDataReady = (dataId, props = null) => {
     switch (dataId) {
+      case 'intro':
+        return intro && <Intro data={intro} bbc={handleBBC} />;
+      case 'indexes':
+        return <IndexesContainer genres={genres} handleGenre={() => console.log('things')} />;
+      case 'story':
+        return <StoriesContainer genres={genres} stories={story} authorsData={authors} />;
+      case 'submit':
+        return <div>SUBMIT</div>;
+      case 'libraryCard':
+        return <LibraryCardContianer data={libraryCard} />;
       case 'author':
         return <AuthorContainer {...props} />;
       case 'stories':
@@ -89,36 +72,38 @@ class ContentContainer extends Component {
         // TODO: make better loading message
         return <span>LOADING</span>;
     }
-  }
+  };
 
-  handleBBC(text, key) {
+  const handleBBC = (text, key) => {
     return bcc(text, key);
-  }
+  };
 
-  render() {
-    return (
-      <div className="contentContainer left">
+  return (
+    <div className="bodyContainer cell-20">
+      {
+        error && <ErrorHandling error={error} />
+      }
+      <section className="container row wrap">
         {
-          this.props.error && <ErrorHandling error={this.props.error} />
+          allLoaded ?
+            <>
+              <Route exact={true} path="/" render={() => handleDataReady('intro')} />
+              <Route exact={true} path="/stories" render={() => handleDataReady('story')} />
+              <Route exact={true} path="/indexes" render={() => handleDataReady('indexes')} />
+              <Route exact={true} path="/submit" render={() => handleDataReady('submit')} />
+              <Route exact={true} path="/library-card" render={() => handleDataReady('libraryCard')} />
+              <Route exact={true} path="/author/:id" render={(props) => handleDataReady('author', props)} />
+              <Route exact={true} path="/stories/:genre/:id" render={(props) => handleDataReady('stories', props)} />
+            </> :
+            <p>LOADING</p>
         }
-        <section>
-          <Route exact={true} path="/" render={() => this.handleDataReady('intro')} />
-          <Route exact={true} path="/stories" render={() => this.handleDataReady('story')} />
-          <Route exact={true} path="/indexes" render={() => this.handleDataReady('indexes')} />
-          <Route exact={true} path="/submit" render={() => this.handleDataReady('submit')} />
-          <Route exact={true} path="/library-card" render={() => this.handleDataReady('libraryCard')} />
-
-          <Route exact={true} path="/author/:id" render={(props) => this.handleDataReady('author', props)} />
-
-          <Route exact={true} path="/stories/:genre/:id" render={(props) => this.handleDataReady('stories', props)} />
-        </section>
-      </div>
-    );
-  }
-}
+      </section>
+    </div>
+  );
+};
 
 ContentContainer.propType = {
   error: PropTypes.object.isRequired
-}
+};
 
 export default ContentContainer;
