@@ -1,68 +1,46 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from 'react';
+// import PropTypes from 'prop-types';
 import UserName from './UserName';
 import Password from './Password';
 
-class LibraryCardContainer extends Component {
-  constructor() {
-    super();
+const LibraryCardContainer = () => {
+  const [userName, setUserName] = useState('');
+  const [password, setPassword] = useState('');
+  const [showPassword, setShowPassword] = useState(false);
 
-    this.state = {
-      userName: '',
-      password: '',
-      showPassword: false
-    }
-
-    this.submitLibraryCard = this.submitLibraryCard.bind(this);
-    this.handleUserNameChange = this.handleUserNameChange.bind(this);
-    this.handlePasswordChange = this.handlePasswordChange.bind(this);
-  }
-
-  submitLibraryCard (event) {
+  const submitLibraryCard = event => {
     event.preventDefault();
 
-    if (!this.state.showPassword) {
-      this.setState(prevState => ({
-        ...prevState,
-        showPassword: true
-      }));
+    if (!showPassword) {
+      setShowPassword(true);
 
       return;
     }
 
-    console.log(this.state.userName + ' ' + this.state.password);
-  }
+    console.log(userName + ' ' + password);
+  };
 
-  handleUserNameChange(event) {
-    this.setState({ userName: event.target.value});
-  }
+  const handleUserNameChange = event => {
+    setUserName(event.target.value);
+  };
 
-  handlePasswordChange(event) {
-    this.setState({ password: event.target.value });
-  }
+  const handlePasswordChange = event => {
+    setPassword(event.target.value);
+  };
 
-
-  render() {
-    const {
-      userName,
-      password,
-      showPassword
-    } = this.state;
-
-    return (
-      <form onSubmit={this.submitLibraryCard}>
-        <label>
-          Library Card
-        </label>
-        {
-          !showPassword && <UserName userName={userName} handleUserNameChange={this.handleUserNameChange} />
-        }
-        {
-          userName && showPassword && <Password password={password} handlePasswordChange={this.handlePasswordChange} />
-        }
-      </form>
-    );
-  }
-}
+  return (
+    <form onSubmit={submitLibraryCard}>
+      <label>
+        Library Card
+      </label>
+      {
+        !showPassword && <UserName userName={userName} handleUserNameChange={handleUserNameChange} />
+      }
+      {
+        userName && showPassword && <Password password={password} handlePasswordChange={handlePasswordChange} />
+      }
+    </form>
+  );
+};
 
 export default LibraryCardContainer;
