@@ -1,81 +1,67 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 // import PropTypes from 'prop-types';
-import './Authors.css';
+import './Authors.scss';
 import AuthorInitial from './AuthorInitial';
 import AuthorByInitial from './AuthorByInitial';
 
-class AuthorsContainer extends Component {
-  constructor(props) {
-    super(props);
+const AuthorsContainer = ({data}) => {
+  const [selectedInitial, setSelectedInitial] = useState(null);
+  // const [authors, setAuthors] = useState(data.letters);
+  const authors = data.letters;
+  // const [allAuthors, setAllAuthors] = useState(null);
+  // const [activeInitials, setActiveInitials] = useState(data.active);
+  const activeInitials = data.active;
 
-    this.state = {
-      selectedInitial: null,
-      authors: props.authors.letters,
-      activeInitials: props.authors.active
-    };
+  const handleInitialClick = id => {
+    setSelectedInitial(id);
+    // setAllAuthors(null);
+  };
 
-    this.buildList = this.buildList.bind(this);
-    this.handleInitialClick = this.handleInitialClick.bind(this);
-  }
-
-  handleInitialClick (id) {
-    this.setState(prevState => ({
-      ...prevState,
-      selectedInitial: id,
-      allAuthors: null
-    }))
-  }
-
-  buildList () {
+  const buildList = () => {
+    const data = [];
+    const classname = ''; //'cell-6';
     let counter = 26;
-    let data = [];
 
     while (counter --) {
       const letter = String.fromCharCode(65 + counter);
 
-      if (this.state.activeInitials.indexOf(letter) >= 0) {
+      if (activeInitials.indexOf(letter) >= 0) {
         data.unshift((
-          <li key={letter} onClick={() => (this.handleInitialClick(letter))}>{letter}</li>
+          <li key={letter} className={classname} onClick={() => (handleInitialClick(letter))}>{letter}</li>
         ));
       } else {
         data.unshift((
-          <li key={letter}>-{letter}-</li>
+          <li key={letter} className={classname}>-{letter}-</li>
         ));
       }
     }
 
-    if (this.state.activeInitials.indexOf('NUMBER') >= 0) {
-      data.unshift(<li key="NUMBER" onClick={() => (this.handleInitialClick('NUMBER'))}>#</li>);
+    if (activeInitials.indexOf('NUMBER') >= 0) {
+      data.unshift(<li key="NUMBER" className={classname} onClick={() => (handleInitialClick('NUMBER'))}>#</li>);
     } else {
-      data.unshift(<li key="NUMBER">-#-</li>);
+      data.unshift(<li key="NUMBER" className={classname}>-#-</li>);
     }
 
     return data;
-  }
+  };
 
-  getAuthors (initial) {
-    if (this.state.authors) {
-      return this.state.authors[initial];
-    }
+  const getAuthors = initial => {
+    if (authors) return authors[initial];
+
     return null;
-  }
+  };
 
-  render () {
-    const {
-      selectedInitial
-    } = this.state;
-    return (
-      <div>
-        <h2>Authors:</h2>
-        <AuthorInitial buildList={this.buildList} />
-        <div>
-          {
-            selectedInitial && <AuthorByInitial authors={this.getAuthors(selectedInitial)} initial={selectedInitial} />
-          }
-        </div>
+  return (
+    <div className="">
+      <h2>Authors:</h2>
+      <AuthorInitial buildList={buildList} />
+      <div className="">
+        {
+          selectedInitial && <AuthorByInitial authors={getAuthors(selectedInitial)} initial={selectedInitial} />
+        }
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 export default AuthorsContainer;
