@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React  from 'react';
 import PropTypes from 'prop-types';
 
 const Password = ({
@@ -11,11 +11,11 @@ const Password = ({
       <input type="submit" value="Submit" />
     </div>
   );
-}
+};
 
 Password.propTypes = {
   password: PropTypes.string,
   handlePasswordChange: PropTypes.func.isRequired
-}
+};
 
 export default Password;
