@@ -4,29 +4,29 @@ import { Link } from 'react-router-dom';
 import { simpleString } from '../../../Utility/Utilities';
 
 const AuthorByInitial = ({ authors, initial }) => {
-  function setAuthors () {
+  const setAuthors = () => {
     if (authors === null) {
       const text = (initial === 'NUMBER') ? 'a number' : `the letter ${initial}`;
-      return (<li key={initial}>No authors starting with {text}</li>);
+      return (<li key={initial} className="">No authors starting with {text}</li>);
     }
 
     return authors.map(author => (
-      <li key={author.guid}>
+      <li key={author.guid} className="">
         <Link to={'./author/' + simpleString(author.name)}>{author.name}</Link>
       </li>
     ));
-  }
+  };
 
   return (
-    <ul>
+    <ul className="authorInitials menu">
       {setAuthors()}
     </ul>
   );
-}
+};
 
 AuthorByInitial.propTypes = {
   authors: PropTypes.array,
   initial: PropTypes.string.isRequired
-}
+};
 
 export default AuthorByInitial;
