@@ -3,15 +3,15 @@ import PropTypes from 'prop-types';
 
 const AuthorInitial = ({ buildList }) => {
   return (
-    <ul>
+    <ul id="initialList" className="menu">
       { buildList() }
     </ul>
   );
-}
+};
 
 AuthorInitial.propTypes = {
   buildList: PropTypes.func.isRequired
-}
+};
 
 
 export default AuthorInitial;
