@@ -1,33 +1,27 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React  from 'react';
+import './Genres.scss'
+// import PropTypes from 'prop-types';
 
-class GenresContainer extends Component {
-  render () {
-    const {
-      genres,
-      handleGenre
-    } = this.props;
+const GenresContainer = ({genres, handleGenre}) => {
+  return (
+    <div className="genresContainer">
+      <h2>Genres:</h2>
+      <ul className="menu">
+        {
+          genres.map(genre => (
+            <li className="genre" key={genre} onClick={() => handleGenre(genre)}>{genre}</li>
+          ))
+        }
+      </ul>
+    </div>
+  );
+};
 
-    return (
-      <div>
-        <h2>Genres:</h2>
-        <ul>
-          {
-            genres.map(genre => (
-              <li key={genre} onClick={() => handleGenre(genre)}>{genre}</li>
-            ))
-          }
-        </ul>
-      </div>
-    );
-  }
-}
-
-GenresContainer.propTypes = {
-  genres: PropTypes.arrayOf(
-    PropTypes.string.isRequired
-  ).isRequired,
-  handleGenre: PropTypes.func.isRequired
-}
+// GenresContainer.propTypes = {
+//   genres: PropTypes.arrayOf(
+//     PropTypes.string.isRequired
+//   ).isRequired,
+//   handleGenre: PropTypes.func.isRequired
+// };
 
 export default GenresContainer;
