@@ -1,14 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-export function simpleString (text) {
+export const simpleString = text => {
   return text.toLowerCase().replace(/[^a-z0-9]/g, "-");
-}
+};
 
-export function bcc (text, key) {
-  if (!text) {
-    return text;
-  }
+export const bcc = (text, key) => {
+  if (!text) return text;
+
   // text = text.replace(/\[color:(.+)\](.+)\[\/color\]/gi, "<span style='color:$1;'>$2</span>");
   // text = text.replace(/\[(b|i|q)\](.+?)\[\/(b|i|q)\]/gi, "<$1>$2</$3>");
   // text = text.replace(/\[(gnome|smilie|winkie|happie)\]/gi, "<img src='" + imageUrls.design + "$1.png' alt='$1' />");
@@ -17,13 +16,11 @@ export function bcc (text, key) {
   const splitOn = /\[link \((.+?)\)\](.+)\[\/link\]/;
   const split = text.split(splitOn);
 
-  if (split.length === 1) {
-    return <p key={key}>{text}</p>;
-  }
+  if (split.length === 1) return <p key={key}>{text}</p>;
 
   let data = [];
 
-  function forEach () {
+  const forEach = () => {
     data = [];
 
     // This is not the best needs a refactor, what happens when it's not a link or the link is the first item?
@@ -41,4 +38,4 @@ export function bcc (text, key) {
   };
 
   return forEach();
-}
+};
